chore(events): split workspace app validators from catalog
Keep eventstream catalog.go under the 800-line revive limit so pre-push lint:go can pass on release/0704.
